date selector added, not actually used for anything yet though
before it's used we need to figure out for real if we're going to use
naive dates or proper chrono dates. unsure at the moment. naivedates
seem better for days, maybe? idk it's hard, i need to figure out how
that conversion works a little better than i understand it at the
moment.